cross-role-translator-server/requirements.txt
fastapi
uvicorn[standard]
langgraph
langchain
openai
python-dotenv
typing_extensions
volcengine-python-sdk